requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==3.20.1
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0

bson~=0.5.10
google~=3.0.0